add logic to check for file diffs, rename validation utils module
